Patch by Ulrich Arnold which fixes some issues which occurred due to wrong mergin by me, some fixed casts and some fixes for Windoes buils.
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@1165 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfContentsTokenizer.cpp
@@ -199,7 +199,7 @@ bool PdfContentsTokenizer::ReadInlineImgData( EPdfContentsType& reType, const ch
             if (e == 'E' && i == 'I')
             {
                 m_buffer.GetBuffer()[counter] = '\0';
-                rVariant = PdfData(m_buffer.GetBuffer(), counter);
+                rVariant = PdfData(m_buffer.GetBuffer(), static_cast<size_t>(counter));
                 reType = ePdfContentsType_ImageData;
                 m_readingInlineImgData = false;
                 return true;
